Impute genetically regulated omics values

Build sparse weight triplets and generate `.grom`, `.gid`, and `.sid` outputs from model weights and PLINK2 genotype inputs.

gromtools_impute(
  weights_table,
  grom_pfx,
  pgen_dir,
  snp_chunk = 1000L,
  sample_subset = integer(0),
  CHUNK = 1000L,
  exportChunk = 256L,
  meanimpute = TRUE,
  is_round = TRUE
)

Arguments

weights_table

A `data.frame` or `data.table` of model weights.

grom_pfx

Output prefix for the generated grom files.

pgen_dir

Path to a directory containing PLINK2 `.pgen`, `.pvar`, and `.psam` files.

snp_chunk

Integer chunk size used while streaming SNPs.

sample_subset

Optional integer vector of 1-based sample indices to retain.

CHUNK

Integer chunk size passed to the low-level imputation engine.

exportChunk

Integer chunk size used when exporting results.

meanimpute

Logical indicating whether missing dosages should be mean imputed.

is_round

Logical indicating whether mean imputation should use rounded means.

Value

Invisibly returns the result of the imputation pipeline after writing output files to disk.

Examples

pgen_dir <- system.file(
  "extdata",
  "synthetic_chromosomes",
  package = "gromtools"
)
db_directory <- system.file(
  "extdata",
  "synth_small_variant_weights_db",
  package = "gromtools"
)
model_weights_table <- read_db_dir(db_dir = db_directory)

out_dir <- file.path(tempdir(), "tmp_grom_run")
dir.create(out_dir, recursive = TRUE, showWarnings = FALSE)
grom_pfx <- file.path(out_dir, "synth_example")

gromtools_impute(
  weights_table = model_weights_table,
  grom_pfx = grom_pfx,
  pgen_dir = pgen_dir
)
#> ### impute_grom() started at: 2026-04-15 14:22:12.560644
#> All chromosome tags were uniquely mapped to a .pgen file.
#> Beginning calculations for chromosome-tag: chr1...
#> Total time for chromosome-tag chr1: 0.0789432525634766
#> Beginning calculations for chromosome-tag: chr10...
#> Total time for chromosome-tag chr10: 0.0019385814666748
#> Beginning calculations for chromosome-tag: chr11...
#> Total time for chromosome-tag chr11: 0.0281522274017334
#> Beginning calculations for chromosome-tag: chr12...
#> Total time for chromosome-tag chr12: 0.062267541885376
#> Beginning calculations for chromosome-tag: chr13...
#> Total time for chromosome-tag chr13: 0.00835847854614258
#> Beginning calculations for chromosome-tag: chr14...
#> Total time for chromosome-tag chr14: 0.0035245418548584
#> Beginning calculations for chromosome-tag: chr15...
#> Total time for chromosome-tag chr15: 0.00145053863525391
#> Beginning calculations for chromosome-tag: chr16...
#> Total time for chromosome-tag chr16: 0.00172924995422363
#> Beginning calculations for chromosome-tag: chr17...
#> Total time for chromosome-tag chr17: 0.00353789329528809
#> Beginning calculations for chromosome-tag: chr18...
#> Total time for chromosome-tag chr18: 0.000724554061889648
#> Beginning calculations for chromosome-tag: chr19...
#> Total time for chromosome-tag chr19: 0.0028538703918457
#> Beginning calculations for chromosome-tag: chr2...
#> Total time for chromosome-tag chr2: 0.00377988815307617
#> Beginning calculations for chromosome-tag: chr20...
#> Total time for chromosome-tag chr20: 0.000774621963500977
#> Beginning calculations for chromosome-tag: chr21...
#> Total time for chromosome-tag chr21: 0.000919342041015625
#> Beginning calculations for chromosome-tag: chr22...
#> Total time for chromosome-tag chr22: 0.0011899471282959
#> Beginning calculations for chromosome-tag: chr3...
#> Total time for chromosome-tag chr3: 0.0101902484893799
#> Beginning calculations for chromosome-tag: chr4...
#> Total time for chromosome-tag chr4: 0.00209188461303711
#> Beginning calculations for chromosome-tag: chr5...
#> Total time for chromosome-tag chr5: 0.00165843963623047
#> Beginning calculations for chromosome-tag: chr6...
#> Total time for chromosome-tag chr6: 0.00206327438354492
#> Beginning calculations for chromosome-tag: chr7...
#> Total time for chromosome-tag chr7: 0.00374579429626465
#> Beginning calculations for chromosome-tag: chr8...
#> Total time for chromosome-tag chr8: 0.00214624404907227
#> Beginning calculations for chromosome-tag: chr9...
#> Total time for chromosome-tag chr9: 0.0107526779174805
#> Chromosome tags detected in weights table:
#> chr1, chr10, chr11, chr12, chr13, chr14, chr15, chr16, chr17, chr18, chr19, chr2, chr20, chr21, chr22, chr3, chr4, chr5, chr6, chr7, chr8, chr9
#> All chromosome tags of column 'chromosome' were mapped successfully.
#> All .pgen files from directory '/tmp/RtmpVJIcNJ/temp_libpathb66c527d5ad8/gromtools/extdata/synthetic_chromosomes' were successfully mapped.
#> impute_grom() completed. Total time: 0.244132280349731