Read selected values from a grom output

Stream selected model, gene, and sample combinations from an existing grom output on disk.

gromtools_read(grom_pfx, models = NULL, genes = NULL, samples = NULL)

Arguments

grom_pfx

Path prefix to an existing grom output without the file extension.

models

Optional character vector of model IDs to retain. `NULL` selects all models.

genes

Optional character vector of gene IDs to retain. `NULL` selects all genes.

samples

Optional character vector of sample IDs to retain. `NULL` selects all samples.

Value

A numeric matrix with samples in rows and selected model-gene pairs in columns.

Examples

pgen_dir <- system.file(
  "extdata",
  "synthetic_chromosomes",
  package = "gromtools"
)
db_directory <- system.file(
  "extdata",
  "synth_small_variant_weights_db",
  package = "gromtools"
)
model_weights_table <- read_db_dir(db_dir = db_directory)

out_dir <- file.path(tempdir(), "tmp_grom_run")
dir.create(out_dir, recursive = TRUE, showWarnings = FALSE)
grom_pfx <- file.path(out_dir, "synth_example")

gromtools_impute(
  weights_table = model_weights_table,
  grom_pfx = grom_pfx,
  pgen_dir = pgen_dir
)
#> ### impute_grom() started at: 2026-04-15 14:22:13.597579
#> All chromosome tags were uniquely mapped to a .pgen file.
#> Beginning calculations for chromosome-tag: chr1...
#> Total time for chromosome-tag chr1: 0.00439691543579102
#> Beginning calculations for chromosome-tag: chr10...
#> Total time for chromosome-tag chr10: 0.0014653205871582
#> Beginning calculations for chromosome-tag: chr11...
#> Total time for chromosome-tag chr11: 0.00174212455749512
#> Beginning calculations for chromosome-tag: chr12...
#> Total time for chromosome-tag chr12: 0.00538778305053711
#> Beginning calculations for chromosome-tag: chr13...
#> Total time for chromosome-tag chr13: 0.000756502151489258
#> Beginning calculations for chromosome-tag: chr14...
#> Total time for chromosome-tag chr14: 0.00197243690490723
#> Beginning calculations for chromosome-tag: chr15...
#> Total time for chromosome-tag chr15: 0.00134515762329102
#> Beginning calculations for chromosome-tag: chr16...
#> Total time for chromosome-tag chr16: 0.00135207176208496
#> Beginning calculations for chromosome-tag: chr17...
#> Total time for chromosome-tag chr17: 0.00204682350158691
#> Beginning calculations for chromosome-tag: chr18...
#> Total time for chromosome-tag chr18: 0.000733613967895508
#> Beginning calculations for chromosome-tag: chr19...
#> Total time for chromosome-tag chr19: 0.00211763381958008
#> Beginning calculations for chromosome-tag: chr2...
#> Total time for chromosome-tag chr2: 0.00527191162109375
#> Beginning calculations for chromosome-tag: chr20...
#> Total time for chromosome-tag chr20: 0.000856637954711914
#> Beginning calculations for chromosome-tag: chr21...
#> Total time for chromosome-tag chr21: 0.000834465026855469
#> Beginning calculations for chromosome-tag: chr22...
#> Total time for chromosome-tag chr22: 0.000993490219116211
#> Beginning calculations for chromosome-tag: chr3...
#> Total time for chromosome-tag chr3: 0.00333285331726074
#> Beginning calculations for chromosome-tag: chr4...
#> Total time for chromosome-tag chr4: 0.00223112106323242
#> Beginning calculations for chromosome-tag: chr5...
#> Total time for chromosome-tag chr5: 0.00375771522521973
#> Beginning calculations for chromosome-tag: chr6...
#> Total time for chromosome-tag chr6: 0.00165343284606934
#> Beginning calculations for chromosome-tag: chr7...
#> Total time for chromosome-tag chr7: 0.0018467903137207
#> Beginning calculations for chromosome-tag: chr8...
#> Total time for chromosome-tag chr8: 0.00166487693786621
#> Beginning calculations for chromosome-tag: chr9...
#> Total time for chromosome-tag chr9: 0.002166748046875
#> Chromosome tags detected in weights table:
#> chr1, chr10, chr11, chr12, chr13, chr14, chr15, chr16, chr17, chr18, chr19, chr2, chr20, chr21, chr22, chr3, chr4, chr5, chr6, chr7, chr8, chr9
#> All chromosome tags of column 'chromosome' were mapped successfully.
#> All .pgen files from directory '/tmp/RtmpVJIcNJ/temp_libpathb66c527d5ad8/gromtools/extdata/synthetic_chromosomes' were successfully mapped.
#> impute_grom() completed. Total time: 0.0547688007354736

grom_mat <- gromtools_read(
  grom_pfx = grom_pfx,
  models = c("AMR_subclass_IN_SST", "AMR_subclass_VLMC"),
  genes = c("ENSG00000003987", "ENSG00000053900"),
  samples = c(
    "sample0001", "sample0014", "sample0034", "sample0039", "sample0043"
  )
)

print(grom_mat)
#>            AMR_subclass_IN_SST_ENSG00000003987
#> sample0001                           0.3191238
#> sample0014                           0.4542425
#> sample0034                           0.4138578
#> sample0039                           0.2676161
#> sample0043                           0.3721308
#>            AMR_subclass_IN_SST_ENSG00000053900
#> sample0001                           0.2737647
#> sample0014                          -0.4481538
#> sample0034                          -0.2212396
#> sample0039                          -0.3217697
#> sample0043                           0.2609762
#>            AMR_subclass_VLMC_ENSG00000053900
#> sample0001                        -0.5058091
#> sample0014                        -0.9249124
#> sample0034                        -0.5259653
#> sample0039                        -0.9044481
#> sample0043                        -0.6188168