Impute genetically regulated omics values

Build sparse weight triplets and generate `.grom`, `.gid`, and `.sid` outputs from model weights and PLINK2 genotype inputs.

Usage

grom_impute(
  weights_table,
  grom_pfx,
  pgen_dir,
  snp_chunk = 1000L,
  sample_subset = integer(0),
  CHUNK = 1000L,
  exportChunk = 256L,
  meanimpute = TRUE,
  is_round = TRUE
)

Arguments

weights_table

A `data.frame` or `data.table` of model weights.

grom_pfx

Output prefix for the generated grom files.

pgen_dir

Path to a directory containing PLINK2 `.pgen`, `.pvar`, and `.psam` files.

snp_chunk

Integer chunk size used while streaming SNPs. Specifies how many genotype records are loaded at a time and can be used to control peak and average memory usage.

sample_subset

Integer vector specifying which samples to load and use for imputation.

CHUNK

Integer chunk size passed to the low-level imputation engine. Specifies the tile size used during computation and can help maintain efficient CPU-cache utilization.

exportChunk

Integer chunk size used when exporting results. Specifies how many output columns are written at a time and can be used to control peak and average memory usage.

meanimpute

Logical indicating whether missing dosages should be mean imputed.

is_round

Logical indicating whether mean imputation should use rounded means.

Value

Invisibly returns the result of the imputation pipeline after writing output files to disk.

Examples

pgen_dir <- system.file(
  "extdata",
  "synthetic_chromosomes",
  package = "gromtools"
)
db_directory <- system.file(
  "extdata",
  "synth_small_variant_weights_db",
  package = "gromtools"
)
model_weights_table <- read_db_dir(db_dir = db_directory)

out_dir <- file.path(tempdir(), "tmp_grom_run")
dir.create(out_dir, recursive = TRUE, showWarnings = FALSE)
grom_pfx <- file.path(out_dir, "synth_example")

grom_impute(
  weights_table = model_weights_table,
  grom_pfx = grom_pfx,
  pgen_dir = pgen_dir
)
#> ### impute_grom() started at: 2026-04-15 17:44:16.229277
#> All chromosome tags were uniquely mapped to a .pgen file.
#> Beginning calculations for chromosome-tag: chr1...
#> Total time for chromosome-tag chr1: 0.0103895664215088
#> Beginning calculations for chromosome-tag: chr10...
#> Total time for chromosome-tag chr10: 0.00278258323669434
#> Beginning calculations for chromosome-tag: chr11...
#> Total time for chromosome-tag chr11: 0.00250816345214844
#> Beginning calculations for chromosome-tag: chr12...
#> Total time for chromosome-tag chr12: 0.00388383865356445
#> Beginning calculations for chromosome-tag: chr13...
#> Total time for chromosome-tag chr13: 0.000916719436645508
#> Beginning calculations for chromosome-tag: chr14...
#> Total time for chromosome-tag chr14: 0.0039827823638916
#> Beginning calculations for chromosome-tag: chr15...
#> Total time for chromosome-tag chr15: 0.00235319137573242
#> Beginning calculations for chromosome-tag: chr16...
#> Total time for chromosome-tag chr16: 0.00345468521118164
#> Beginning calculations for chromosome-tag: chr17...
#> Total time for chromosome-tag chr17: 0.00486660003662109
#> Beginning calculations for chromosome-tag: chr18...
#> Total time for chromosome-tag chr18: 0.000929594039916992
#> Beginning calculations for chromosome-tag: chr19...
#> Total time for chromosome-tag chr19: 0.00400590896606445
#> Beginning calculations for chromosome-tag: chr2...
#> Total time for chromosome-tag chr2: 0.00523734092712402
#> Beginning calculations for chromosome-tag: chr20...
#> Total time for chromosome-tag chr20: 0.00126409530639648
#> Beginning calculations for chromosome-tag: chr21...
#> Total time for chromosome-tag chr21: 0.00113916397094727
#> Beginning calculations for chromosome-tag: chr22...
#> Total time for chromosome-tag chr22: 0.0015265941619873
#> Beginning calculations for chromosome-tag: chr3...
#> Total time for chromosome-tag chr3: 0.00679659843444824
#> Beginning calculations for chromosome-tag: chr4...
#> Total time for chromosome-tag chr4: 0.00318312644958496
#> Beginning calculations for chromosome-tag: chr5...
#> Total time for chromosome-tag chr5: 0.00297451019287109
#> Beginning calculations for chromosome-tag: chr6...
#> Total time for chromosome-tag chr6: 0.00282001495361328
#> Beginning calculations for chromosome-tag: chr7...
#> Total time for chromosome-tag chr7: 0.00355982780456543
#> Beginning calculations for chromosome-tag: chr8...
#> Total time for chromosome-tag chr8: 0.00386714935302734
#> Beginning calculations for chromosome-tag: chr9...
#> Total time for chromosome-tag chr9: 0.0210702419281006
#> Chromosome tags detected in weights table:
#> chr1, chr10, chr11, chr12, chr13, chr14, chr15, chr16, chr17, chr18, chr19, chr2, chr20, chr21, chr22, chr3, chr4, chr5, chr6, chr7, chr8, chr9
#> All chromosome tags of column 'chromosome' were mapped successfully.
#> All .pgen files from directory '/tmp/RtmpxZmHLX/temp_libpath20d00166cd44c/gromtools/extdata/synthetic_chromosomes' were successfully mapped.
#> impute_grom() completed. Total time: 0.100791215896606