Read selected values from a grom output

Stream selected model, gene, and sample combinations from an existing grom output on disk.

Usage

gromtools_read(grom_pfx, models = NULL, genes = NULL, samples = NULL)

Arguments

grom_pfx

Path prefix to an existing grom output without the file extension.

models

Optional character vector of model IDs to retain. `NULL` selects all models.

genes

Optional character vector of gene IDs to retain. `NULL` selects all genes.

samples

Optional character vector of sample IDs to retain. `NULL` selects all samples.

Value

A numeric matrix with samples in rows and selected model-gene pairs in columns.

Examples

pgen_dir <- system.file(
  "extdata",
  "synthetic_chromosomes",
  package = "gromtools"
)
db_directory <- system.file(
  "extdata",
  "synth_small_variant_weights_db",
  package = "gromtools"
)
model_weights_table <- read_db_dir(db_dir = db_directory)

out_dir <- file.path(tempdir(), "tmp_grom_run")
dir.create(out_dir, recursive = TRUE, showWarnings = FALSE)
grom_pfx <- file.path(out_dir, "synth_example")

gromtools_impute(
  weights_table = model_weights_table,
  grom_pfx = grom_pfx,
  pgen_dir = pgen_dir
)
#> ### impute_grom() started at: 2026-04-15 16:03:49.753819
#> All chromosome tags were uniquely mapped to a .pgen file.
#> Beginning calculations for chromosome-tag: chr1...
#> Total time for chromosome-tag chr1: 0.00989770889282227
#> Beginning calculations for chromosome-tag: chr10...
#> Total time for chromosome-tag chr10: 0.00394845008850098
#> Beginning calculations for chromosome-tag: chr11...
#> Total time for chromosome-tag chr11: 0.00318551063537598
#> Beginning calculations for chromosome-tag: chr12...
#> Total time for chromosome-tag chr12: 0.00498580932617188
#> Beginning calculations for chromosome-tag: chr13...
#> Total time for chromosome-tag chr13: 0.00134706497192383
#> Beginning calculations for chromosome-tag: chr14...
#> Total time for chromosome-tag chr14: 0.00428581237792969
#> Beginning calculations for chromosome-tag: chr15...
#> Total time for chromosome-tag chr15: 0.0136799812316895
#> Beginning calculations for chromosome-tag: chr16...
#> Total time for chromosome-tag chr16: 0.00306391716003418
#> Beginning calculations for chromosome-tag: chr17...
#> Total time for chromosome-tag chr17: 0.00457024574279785
#> Beginning calculations for chromosome-tag: chr18...
#> Total time for chromosome-tag chr18: 0.00130724906921387
#> Beginning calculations for chromosome-tag: chr19...
#> Total time for chromosome-tag chr19: 0.00489497184753418
#> Beginning calculations for chromosome-tag: chr2...
#> Total time for chromosome-tag chr2: 0.00547146797180176
#> Beginning calculations for chromosome-tag: chr20...
#> Total time for chromosome-tag chr20: 0.00147223472595215
#> Beginning calculations for chromosome-tag: chr21...
#> Total time for chromosome-tag chr21: 0.00156521797180176
#> Beginning calculations for chromosome-tag: chr22...
#> Total time for chromosome-tag chr22: 0.00203037261962891
#> Beginning calculations for chromosome-tag: chr3...
#> Total time for chromosome-tag chr3: 0.00928187370300293
#> Beginning calculations for chromosome-tag: chr4...
#> Total time for chromosome-tag chr4: 0.0045475959777832
#> Beginning calculations for chromosome-tag: chr5...
#> Total time for chromosome-tag chr5: 0.00328731536865234
#> Beginning calculations for chromosome-tag: chr6...
#> Total time for chromosome-tag chr6: 0.00389981269836426
#> Beginning calculations for chromosome-tag: chr7...
#> Total time for chromosome-tag chr7: 0.00434637069702148
#> Beginning calculations for chromosome-tag: chr8...
#> Total time for chromosome-tag chr8: 0.0036470890045166
#> Beginning calculations for chromosome-tag: chr9...
#> Total time for chromosome-tag chr9: 0.0106630325317383
#> Chromosome tags detected in weights table:
#> chr1, chr10, chr11, chr12, chr13, chr14, chr15, chr16, chr17, chr18, chr19, chr2, chr20, chr21, chr22, chr3, chr4, chr5, chr6, chr7, chr8, chr9
#> All chromosome tags of column 'chromosome' were mapped successfully.
#> All .pgen files from directory '/tmp/Rtmp7MfgrR/temp_libpath162917a3b04c0/gromtools/extdata/synthetic_chromosomes' were successfully mapped.
#> impute_grom() completed. Total time: 0.115006923675537

grom_mat <- gromtools_read(
  grom_pfx = grom_pfx,
  models = c("AMR_subclass_IN_SST", "AMR_subclass_VLMC"),
  genes = c("ENSG00000003987", "ENSG00000053900"),
  samples = c(
    "sample0001", "sample0014", "sample0034", "sample0039", "sample0043"
  )
)

print(grom_mat)
#>            AMR_subclass_IN_SST_ENSG00000003987
#> sample0001                           0.3191238
#> sample0014                           0.4542425
#> sample0034                           0.4138578
#> sample0039                           0.2676161
#> sample0043                           0.3721308
#>            AMR_subclass_IN_SST_ENSG00000053900
#> sample0001                           0.2737647
#> sample0014                          -0.4481538
#> sample0034                          -0.2212396
#> sample0039                          -0.3217697
#> sample0043                           0.2609762
#>            AMR_subclass_VLMC_ENSG00000053900
#> sample0001                        -0.5058091
#> sample0014                        -0.9249124
#> sample0034                        -0.5259653
#> sample0039                        -0.9044481
#> sample0043                        -0.6188168